requirements.txt
filelock==3.18.0
fsspec==2025.3.2
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
scikit-learn==1.6.1
scipy==1.15.2
shiboken6==6.9.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
